add methods to take picture and safe it
I have implemented dispatchTakePictureIntent() which first creates a new File where the picture will be stored with the help of createImageFile(). Here the File is created with a specific name and location.
After that th camera is opened by an Intent.
Furthermore, I added the onActivityResult() to add the taken picture into the corresponding ImageView.