add qrButton to the HomeScreenActivity
I have added the new qrButton to the HomeScreenActivity with the feature that the button will bounce slightly when it is clicked. Furthermore, i have initialized two transition xmls for the bounce effect. The onTouchListener() must return false so that the onClickListener() is able to perform its task. The user is then navigated to the QR_Reader class.