add bounce.xml
The animation first goes to 0.9x/y of the root size. Then it gets to 1.1x/y of it.
The animation ends after 300 milliseconds and it takes place around the center of x and y. It then stays in this position.
The second animation simply scales the button back to its first size. It starts 200 millisecond after the first and ends in same time.