add QR_reader class
This class sets up and uses the device camera to scan QR codes.It uses CameraX for camera management and ML Kit for QR code scanning.
First in onCreate() the elements of the xml are initialized. It checks whether the app has permission to use the camera or not. If yes, than startCamera() is called, this method is responsible for initializing and starting the camera. When there is no exception, the previewAnalyser() is called. The previewAnalyser() method configures and binds the camera preview and image analysis (for QR code scanning) to the activity's lifecycle, setting up the QR_Analyser and selecting the back camera.