add QR_Analyser class
The QR_Analyser class is an implementation of ImageAnalysis.Analyzer used for processing images from the camera to detect and read QR-Codes. It utilizes Google ML Kit's BarcodeScanning API to perform QR-Code detection.
First a BarcodeScanner is created which only reads QR-Codes. Then analyze() is used to detect any QR-Codes in the Preview/InputImage. As soon as it successful, it will display the text of the QR-Code.